Major Changes.
I have reconfigures most of the write to file using two different
methods of writing to the file. Another change was the fact the the
"pos" array was split into multiple arrays allowing for storing all
data. Public voids were added to be run by buttons in the unity game.
Two things are now a problem including that when I add a new set it
overwrites everything including the lognum line at the first line of the
file and the last issue of issue of:

ndexOutOfRangeException: Array index is out of
range.
Data_Receive_and_Display.loadset () (at
Assets/Data_Receive_and_Display.cs:74)